// frontend/src/api/models/CashAdvanceStatusEnum.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
/**
 * * `pending` - Pending
 * * `approved` - Approved
 * * `rejected` - Rejected
 * * `disbursed` - Disbursed
 * * `repaid` - Repaid
 */
export enum CashAdvanceStatusEnum {
    PENDING = 'pending',
    APPROVED = 'approved',
    REJECTED = 'rejected',
    DISBURSED = 'disbursed',
    REPAID = 'repaid',
}